fix missing methods
